完成衣服编辑功能，开始做搭配相关部分

+++ ZJArmoire/ZJArmoire/Classes/Module/Camera/ViewController/ZJACameraController.swift
@@ -16,7 +16,7 @@ class ZJACameraController: UIViewController {
     var avCaptureSesstion: AVCaptureSession?
     var addPhotoBlock: ConfirmPhotoCallback?
 //    var typeName: String?
-    var yiguiType:NSInteger?
+    var yiguiType:Int?
 
     override func viewDidLoad() {
         super.viewDidLoad()
